src/DCT.ValidationService.POC/BusinessRules.POC/RuleLearnDelFAMType66/ExclusionRules/RestartLearnerDelFamRule.cs
using BusinessRules.POC.Interfaces;
using BusinessRules.POC.Models;
using BusinessRules.POC.ReferenceData;
using DCT.ILR.Model;

namespace BusinessRules.POC.RuleLearnDelFAMType66.ExclusionRules
{
    public class RestartLearnerDelFamRule : ILearnerDelFam66ExclusionRuleForFam
    {
        private readonly string _restartFamType;
        private readonly string _restartFamCode;


        public RestartLearnerDelFamRule(IReferenceData<string, string> referenceData)
        {
            _restartFamType = referenceData.Get(AppConstants.LearnDelFam66RestartFamType);
            _restartFamCode = referenceData.Get(AppConstants.LearnDelFam66RestartFamCode);


        }
        public bool Evaluate(MessageLearnerLearningDeliveryLearningDeliveryFAM ldFam)
        {
            if (ldFam == null) return false;

            return ldFam.LearnDelFAMType == _restartFamType && ldFam.LearnDelFAMCode == _restartFamCode;
        }
    }
}